requirements.txt
google-generativeai==0.7.2
httpx==0.27.0
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
streamlit==1.38.0
streamlit-chat==0.1.1
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
xlsxwriter==3.2.0
sympy==1.13.3
python-dotenv==1.0.1
requests==2.31.0
typing-extensions==4.12.2
pytest==8.3.2
pytest-asyncio==0.24.0